+++ frontend/src/mythicaldash/admin/Eggs.ts
@@ -79,7 +79,7 @@ class Eggs {
         pterodactylEggId: number,
         enabled: string = 'false',
         image_id: number | null,
-        vip: string = 'false',
+        vip_only: string = 'false',
     ) {
         const formData = new FormData();
         formData.append('name', name);
@@ -88,7 +88,7 @@ class Eggs {
         formData.append('pterodactyl_egg_id', pterodactylEggId.toString());
         formData.append('enabled', enabled);
         formData.append('image_id', image_id?.toString() || 'null');
-        formData.append('vip', vip);
+        formData.append('vip_only', vip_only);
         const response = await fetch(`/api/admin/eggs/${id}/update`, {
             method: 'POST',
             body: formData,